requirements.txt
cupy-cuda11x>=12.0.0
torch>=2.0.0
numba>=0.57.0
networkx>=3.1
numpy>=1.24.0
pytest>=7.3.1
cugraph-cu11>=23.02.00
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
jupyterlab>=4.0.0
